refactor: moves MnemonicInput to onboarding components folder
