+++ packages/desktop/views/onboarding/components/MnemonicInput.svelte
