+++ packages/shared/src/components/inputs/index.js
@@ -7,7 +7,6 @@ export { default as Dropdown } from './Dropdown.svelte'
 export { default as Dropdown2 } from './Dropdown2.svelte'
 export { default as Dropzone } from './Dropzone.svelte'
 export { default as IconInput } from './IconInput.svelte'
-export { default as MnemonicInput } from './MnemonicInput.svelte'
 export { default as Input } from './Input.svelte'
 export { default as InputContainer } from './InputContainer.svelte'
 export { default as NetworkInput } from './NetworkInput.svelte'
